+++ mitmproxy/net/tls.py
@@ -3,6 +3,7 @@ import threading
 from collections.abc import Callable
 from collections.abc import Iterable
 from enum import Enum
+from functools import cache
 from functools import lru_cache
 from pathlib import Path
 from typing import Any
@@ -56,6 +57,22 @@ class Verify(Enum):
 DEFAULT_MIN_VERSION = Version.TLS1_2
 DEFAULT_MAX_VERSION = Version.UNBOUNDED
 DEFAULT_OPTIONS = SSL.OP_CIPHER_SERVER_PREFERENCE | SSL.OP_NO_COMPRESSION
+
+
+@cache
+def is_supported_version(version: Version):
+    client_ctx = SSL.Context(SSL.TLS_CLIENT_METHOD)
+    client_ctx.set_min_proto_version(version.value)
+    client_ctx.set_max_proto_version(version.value)
+    client_conn = SSL.Connection(client_ctx)
+    client_conn.set_connect_state()
+
+    try:
+        client_conn.recv(4096)
+    except SSL.WantReadError:
+        return True
+    except SSL.Error:
+        return False
 
 
 class MasterSecretLogger: